+++ core/execution/market_snapshot.go
@@ -161,6 +161,10 @@ func NewMarketFromSnapshot(
 	market.tradableInstrument.Instrument.Product.NotifyOnSettlementPrice(market.settlementPrice)
 	liqEngine.SetGetStaticPricesFunc(market.getBestStaticPricesDecimal)
 
+	if mkt.State == types.MarketStateTradingTerminated {
+		market.tradableInstrument.Instrument.Product.UnsubscribeTradingTerminated(ctx)
+	}
+
 	if mkt.State == types.MarketStateSettled {
 		market.closed = true
 		stateVarEngine.UnregisterStateVariable(asset, mkt.ID)