+++ wallet/api/session_get_chain_id.go
@@ -4,6 +4,7 @@ import (
 	"context"
 
 	"code.vegaprotocol.io/vega/libs/jsonrpc"
+	walletnode "code.vegaprotocol.io/vega/wallet/api/node"
 )
 
 type GetChainIDResult struct {
@@ -11,7 +12,7 @@ type GetChainIDResult struct {
 }
 
 type GetChainID struct {
-	nodeSelector NodeSelector
+	nodeSelector walletnode.Selector
 }
 
 func (h *GetChainID) Handle(ctx context.Context, _ jsonrpc.Params) (jsonrpc.Result, *jsonrpc.ErrorDetails) {
@@ -30,7 +31,7 @@ func (h *GetChainID) Handle(ctx context.Context, _ jsonrpc.Params) (jsonrpc.Resu
 	}, nil
 }
 
-func NewGetChainID(nodeSelector NodeSelector) *GetChainID {
+func NewGetChainID(nodeSelector walletnode.Selector) *GetChainID {
 	return &GetChainID{
 		nodeSelector: nodeSelector,
 	}